+++ Changelog.md
@@ -32,6 +32,7 @@
 - Fix: Sonic Security System settings sometimes do not persist through world reloads
 - Fix: Potential crash in SaltData (ConcurrentModificationException)
 - Fix: Crash when Laser Block/Inventory Scanner ranges are set high
+- Fix: Players are able to mount Security Cameras that have been shut down by an EMP from another mod
 
 --------------------------Changelog for v1.9.9 of SecurityCraft--------------------------